pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-semantic-similarity"
version = "1.0.0"
description = "file: README.md"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "sentence_transformers>=3.1.1",
    "GitPython>=3.1.43",
    "numpy==2.1.0",
    "rank_bm25==0.2.2",
    "nltk==3.9.1",
    "sqlite-vec==0.1.2",
    "click==8.1.7",
]
[project.license]
file = "LICENSE"

[project.scripts]
git-semsim = "src.cli:main"
